[FIX] estaeSalePrice 컬럼 추가

+++ wooreadmin/src/main/java/com/piehouse/wooreadmin/estate/entity/Estate.java
@@ -44,7 +44,7 @@ public class Estate {
     private String estateLongitude;
 
     @Column(nullable = false)
-    private Integer tokenAmount;
+    private Long tokenAmount;
 
     @Column(columnDefinition = "TEXT")
     private String estateDescription;
